--- .eslintrc.js
@@ -1,18 +1,0 @@
-module.exports = {
-  extends: [
-    'next',
-    'next/core-web-vitals',
-    'eslint:recommended',
-    'plugin:react/recommended',
-    'plugin:@typescript-eslint/recommended',
-    'prettier',
-  ],
-  plugins: ['react', '@typescript-eslint'],
-  parserOptions: {
-    ecmaVersion: 2021,
-    sourceType: 'module',
-  },
-  rules: {
-    'react/react-in-jsx-scope': 'off',
-  },
-};
