--- app/projects/layout.tsx
@@ -1,38 +1,0 @@
-import { DocsLayout, type DocsLayoutProps } from 'fumadocs-ui/layouts/docs';
-import type { ReactNode } from 'react';
-import { baseOptions } from '@/app/layout.config';
-import 'fumadocs-twoslash/twoslash.css';
-import { source } from '@/lib/source';
-import React from 'react';
-
-const docsOptions: DocsLayoutProps = {
-  ...baseOptions,
-  tree: source.pageTree,
-  sidebar: {
-    tabs: {
-      transform(option, node) {
-        const meta = source.getNodeMeta(node);
-        if (!meta) return option;
-
-        return {
-          ...option,
-          icon: (
-            <div
-              className="from-fd-background/80 rounded-md border bg-gradient-to-t p-1 shadow-md [&_svg]:size-5"
-              style={{
-                color: `hsl(var(--${meta.file.dirname}-color))`,
-                backgroundColor: `hsl(var(--${meta.file.dirname}-color)/.3)`,
-              }}
-            >
-              {node.icon}
-            </div>
-          ),
-        };
-      },
-    },
-  },
-};
-
-export default function Layout({ children }: { children: ReactNode }) {
-  return <DocsLayout {...docsOptions}>{children}</DocsLayout>;
-}
