--- app/robot.ts
@@ -1,13 +1,0 @@
-import type { MetadataRoute } from 'next';
-import { siteConfig } from '@/site.config';
-
-export default function robots(): MetadataRoute.Robots {
-  return {
-    rules: {
-      userAgent: '*',
-      allow: '/',
-      disallow: '/api/*',
-    },
-    sitemap: `${siteConfig.baseUrl}/sitemap.xml`,
-  };
-}
