--- lib/metadata.ts
@@ -1,7 +1,0 @@
-import { createMetadataImage } from 'fumadocs-core/server';
-import { source } from './source';
-
-export const metadataImage = createMetadataImage({
-  imageRoute: '/projects-og',
-  source,
-});
