+++ package.json
@@ -1,54 +1,46 @@
 {
   "name": "devfolio",
-  "version": "0.0.0",
+  "version": "0.0.1",
   "private": true,
   "scripts": {
     "build": "next build",
-    "dev": "next dev",
+    "dev": "next dev --turbo",
     "start": "next start",
     "postinstall": "fumadocs-mdx",
-    "lint": "eslint .",
+    "lint": "eslint . --ignore-path .gitignore",
     "format": "prettier --check --ignore-path .gitignore .",
     "format:fix": "prettier --write --ignore-path .gitignore .",
     "build:docs": "node ./scripts/generate-docs.mjs"
   },
   "dependencies": {
-    "@fumadocs/cli": "^0.1.1",
     "@radix-ui/react-slot": "^1.2.0",
-    "@scalar/api-client-react": "^1.2.14",
-    "@vercel/analytics": "^1.5.0",
-    "@vercel/speed-insights": "^1.2.0",
-    "ajv": "^8.17.1",
     "class-variance-authority": "^0.7.1",
+    "clsx": "^2.1.1",
     "fumadocs-core": "15.2.5",
     "fumadocs-mdx": "11.5.8",
-    "fumadocs-openapi": "^7.0.13",
     "fumadocs-twoslash": "^3.1.0",
-    "fumadocs-typescript": "^4.0.2",
     "fumadocs-ui": "15.2.5",
-    "next": "15.2.5",
-    "prettier": "^3.5.3",
+    "lucide-react": "^0.487.0",
+    "next": "15.2.4",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
-    "rimraf": "^6.0.1",
     "tailwind-merge": "^3.2.0",
-    "tailwindcss": "^4.1.3",
-    "tailwindcss-animate": "^1.0.7"
+    "tailwindcss-animate": "^1.0.7",
+    "tw-animate-css": "^1.2.5",
+    "twoslash": "^0.3.1"
   },
   "devDependencies": {
-    "eslint-config-prettier": "^10.1.1",
-    "eslint-plugin-prettier": "^5.2.6",
-    "eslint-plugin-react": "^7.37.5",
-    "prettier-plugin-tailwindcss": "^0.6.11",
+    "@tailwindcss/postcss": "^4.1.3",
+    "@types/mdx": "^2.0.13",
     "@types/node": "22.14.0",
     "@types/react": "^19.1.0",
     "@types/react-dom": "^19.1.1",
-    "typescript": "^5.8.3",
-    "@types/mdx": "^2.0.13",
-    "@tailwindcss/postcss": "^4.1.3",
+    "eslint": "^8",
+    "eslint-config-next": "15.2.4",
     "postcss": "^8.5.3",
-    "eslint": "^9.24.0",
-    "eslint-config-next": "15.2.5"
+    "prettier": "3.5.3",
+    "tailwindcss": "^4.1.3",
+    "typescript": "^5.8.3"
   },
   "license": "MIT",
   "repository": {
