--- scripts/generate-docs.mjs
@@ -1,18 +1,0 @@
-import * as OpenAPI from 'fumadocs-openapi';
-import { rimrafSync } from 'rimraf';
-
-const out = './content/projects/(api)';
-
-// clean generated files
-rimrafSync(out, {
-  filter(v) {
-    return !v.endsWith('index.mdx') && !v.endsWith('meta.json');
-  },
-});
-
-void OpenAPI.generateFiles({
-  // input files
-  input: ['./openapi.json'],
-  output: out,
-  groupBy: 'tag',
-});
