+++ source.config.ts
@@ -1,5 +1,6 @@
 import { defineDocs, defineConfig } from 'fumadocs-mdx/config';
-
+import { transformerTwoslash } from 'fumadocs-twoslash';
+import { rehypeCodeDefaultOptions } from 'fumadocs-core/mdx-plugins';
 export const docs = defineDocs({
   dir: 'content/projects',
 });
@@ -11,6 +12,10 @@ export default defineConfig({
         light: 'github-light',
         dark: 'github-dark',
       },
+      transformers: [
+        ...(rehypeCodeDefaultOptions.transformers ?? []),
+        transformerTwoslash(),
+      ],
     },
   },
 });
