+++ src/lib/metadata.ts
@@ -1,0 +1,7 @@
+import { createMetadataImage } from 'fumadocs-core/server';
+import { source } from '@/lib/source';
+
+export const metadataImage = createMetadataImage({
+  imageRoute: '/projects-og',
+  source,
+});
