+++ src/lib/source.ts
@@ -1,0 +1,9 @@
+import { docs } from '@/.source';
+import { loader } from 'fumadocs-core/source';
+
+// `loader()` also assign a URL to your pages
+// See https://fumadocs.vercel.app/docs/headless/source-api for more info
+export const source = loader({
+  baseUrl: '/projects',
+  source: docs.toFumadocsSource(),
+});
